Better styling in AI components #951
